Update Intents
Added additional utterances to the built-in intents.
Converted the Language slot to a custom one. 
Since we currently have only 2 languages, I think it will be easier to check for the few languages we will support, rather than providing error handling for all the other possible language inputs a user may try.